[level 1] Title: 콜라 문제, Time: 0.07 ms, Memory: 16.3 MB -BaekjoonHub

+++ 프로그래머스/1/132267. 콜라 문제/콜라 문제.swift
@@ -7,7 +7,7 @@ func solution(_ a:Int, _ b:Int, _ n:Int) -> Int {
     var bottles = n
     
     while bottles >= a {
-        let tmp = Int(bottles / a) * b
+        let tmp = bottles / a * b
         cnt += tmp
         bottles = bottles % a + tmp
     }
